Refactor `edit_markdown()` and `edit_list()` admin script functions
